homeassistant_gitops/pyproject.toml
[project]
name = "ha-gitops-bridge"
version = "0.1.0"
description = "Home Assistant GitOps Bridge add-on"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "httpx==0.27.2",
  "pydantic==2.9.2",
  "pyyaml==6.0.2",
  "uvicorn==0.31.1",
  "watchdog==5.0.3",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.2",
]
